refactor: code pattern and meta config
