--- src/meta/site.ts
@@ -1,13 +1,0 @@
-export const siteConfig = {
-  name: "NextJs Launchpad",
-  description: "NextJs Scaffolding by CreoWis",
-  themeColor: [
-    { media: "(prefers-color-scheme: light)", color: "white" },
-    { media: "(prefers-color-scheme: dark)", color: "black" },
-  ],
-  icons: {
-    icon: "/favicon.ico",
-    shortcut: "/favicon-16x16.png",
-    apple: "/apple-touch-icon.png",
-  },
-};
