+++ src/modules/home/(sections)/event/index.tsx
